+++ src/helpers/workers/file-loader.d.ts
@@ -3,7 +3,7 @@ declare module "file-loader?name=[name].js!*" {
   export = value;
 }
 
-declare module "worker-loader!*" {
+declare module "worker-loader?name=simulation.worker.js*" {
   class WebpackWorker extends Worker {
     constructor();
   }